trading_bot/requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
colorama>=0.4.6
python-dotenv>=1.0.0
pytz>=2024.1
jinja2>=3.1.0
